MaxBlog/requirements.txt
Django==2.0.5
ipdb==0.10.2
ipython==5.3.0
coverage==4.3.4
coreapi-cli==1.0.6
django-debug-toolbar==1.9.1
django-extensions==2.0.7
Pillow==5.1.0
django-braces==1.13.0
django-rest-swagger==2.2.0
djangorestframework==3.8.2
psycopg2-binary==2.7.4
django-filter==2.0.0
django-nose==1.4.5
factory-boy==2.11.1
celery==4.2.1
redis==2.10.6
flower==0.9.2
beautifulsoup4==4.6.1
requests==2.2.1
channels==1.1.8
daphne==1.4.2
asgi-redis==1.4.3
